indexer

+++ nd4j-backends/nd4j-backend-impls/nd4j-cuda-7.5/src/main/java/org/nd4j/jita/allocator/impl/AllocationShape.java
@@ -10,13 +10,13 @@ import org.nd4j.linalg.api.buffer.DataBuffer;
 @Data
 @NoArgsConstructor
 public class AllocationShape {
-    private int offset = 0;
-    private int length = 0;
+    private long offset = 0;
+    private long length = 0;
     private int stride = 1;
     private int elementSize = 0;
     private DataBuffer.Type dataType = DataBuffer.Type.FLOAT;
 
-    public AllocationShape(int length, int elementSize) {
+    public AllocationShape(long length, int elementSize) {
         this.length = length;
         this.elementSize = elementSize;
     }
